+++ README.md
@@ -1,7 +1,7 @@
 # EgoPhoto
 
 [![License: MIT](https://img.shields.io/badge/License-MIT-blue.svg)](https://opensource.org/licenses/MIT)
-[![CI](https://github.com/vinymeuh/EgoPhoto/workflows/CI/badge.svg)
+![CI](https://github.com/vinymeuh/EgoPhoto/workflows/CI/badge.svg)
 
 Manage my photos my way
