+++ egophoto/images_selector/directory_browser_tab.py
