+++ egophoto/widgets/__init__.py
@@ -1,9 +1,7 @@
-from egophoto.widgets.images_selector import ImagesSelector
 from egophoto.widgets.images_grid import ImagesGrid
 from egophoto.widgets.img_tag_viewer import ImgTagViewer
 
 __all__ = [
-    ImagesSelector,
     ImagesGrid,
     ImgTagViewer,
 ]
